Cherry pick -> ESM Script base class #6367
